flipfile-box/converter.py
# -*- coding: utf-8 -*-
# 真实文件转换模块
# 需要安装相应的转换库

import os
import subprocess
from pathlib import Path

class FileConverter:
    """文件转换器 - 支持真实格式转换"""
    
    def __init__(self):
        self.supported_formats = {
            'document': ['docx', 'doc', 'pdf', 'txt', 'rtf', 'odt'],
            'image': ['jpg', 'jpeg', 'png', 'gif', 'webp', 'bmp', 'tiff'],
            'video': ['mp4', 'avi', 'mov', 'mkv', 'wmv', 'flv'],
            'audio': ['mp3', 'wav', 'flac', 'aac', 'ogg']
        }
    
    def convert(self, input_path, output_format, quality=90):
        """
        转换文件格式
        
        Args:
            input_path: 输入文件路径
            output_format: 目标格式
            quality: 质量（1-100）
            
        Returns:
            转换结果字典
        """
        try:
            input_path = Path(input_path)
            if not input_path.exists():
                return {'success': False, 'error': '文件不存在'}
            
            # 检测文件类型
            file_type = self.detect_file_type(input_path.suffix[1:])
            if not file_type:
                return {'success': False, 'error': '不支持的文件格式'}
            
            # 根据类型选择转换方法
            if file_type == 'document':
                return self.convert_document(input_path, output_format)
            elif file_type == 'image':
                return self.convert_image(input_path, output_format, quality)
            elif file_type == 'video':
                return self.convert_video(input_path, output_format)
            else:
                return {'success': False, 'error': '暂不支持此类型转换'}
                
        except Exception as e:
            return {'success': False, 'error': str(e)}
    
    def detect_file_type(self, ext):
        """检测文件类型"""
        ext = ext.lower()
        for file_type, extensions in self.supported_formats.items():
            if ext in extensions:
                return file_type
        return None
    
    def convert_document(self, input_path, output_format):
        """文档转换"""
        try:
            # 尝试使用 LibreOffice（如果安装）
            output_path = input_path.with_suffix(f'.{output_format}')
            
            # 检查 LibreOffice
            libreoffice_cmd = self.find_libreoffice()
            if libreoffice_cmd:
                subprocess.run([
                    libreoffice_cmd,
                    '--headless',
                    '--convert-to', output_format,
                    str(input_path),
                    '--outdir', str(input_path.parent)
                ], check=True, timeout=60)
                
                return {
                    'success': True,
                    'output': str(output_path),
                    'size': output_path.stat().st_size
                }
            
            # 如果没有 LibreOffice，使用简化方案
            return self.simple_convert(input_path, output_format)
            
        except subprocess.TimeoutExpired:
            return {'success': False, 'error': '转换超时'}
        except Exception as e:
            return {'success': False, 'error': str(e)}
    
    def convert_image(self, input_path, output_format, quality=90):
        """图片转换"""
        try:
            # 尝试使用 Pillow
            from PIL import Image
            
            output_path = input_path.with_suffix(f'.{output_format}')
            
            img = Image.open(input_path)
            
            # 处理 RGBA 模式（PNG 转 JPG 时需要）
            if output_format.lower() in ['jpg', 'jpeg'] and img.mode == 'RGBA':
                img = img.convert('RGB')
            
            # 保存
            save_params = {'quality': quality} if output_format.lower() in ['jpg', 'jpeg', 'webp'] else {}
            img.save(output_path, **save_params)
            
            return {
                'success': True,
                'output': str(output_path),
                'size': output_path.stat().st_size,
                'format': output_format.upper()
            }
            
        except ImportError:
            return {'success': False, 'error': '需要安装 Pillow 库：pip install Pillow'}
        except Exception as e:
            return {'success': False, 'error': str(e)}
    
    def convert_video(self, input_path, output_format):
        """视频转换"""
        try:
            # 尝试使用 FFmpeg
            ffmpeg_cmd = self.find_ffmpeg()
            if not ffmpeg_cmd:
                return {'success': False, 'error': '需要安装 FFmpeg'}
            
            output_path = input_path.with_suffix(f'.{output_format}')
            
            subprocess.run([
                ffmpeg_cmd,
                '-i', str(input_path),
                '-c:v', 'libx264',
                '-c:a', 'aac',
                str(output_path)
            ], check=True, timeout=300)
            
            return {
                'success': True,
                'output': str(output_path),
                'size': output_path.stat().st_size
            }
            
        except subprocess.TimeoutExpired:
            return {'success': False, 'error': '转换超时'}
        except Exception as e:
            return {'success': False, 'error': str(e)}
    
    def simple_convert(self, input_path, output_format):
        """简化转换（复制文件，用于演示）"""
        import shutil
        
        output_path = input_path.with_suffix(f'.{output_format}')
        shutil.copy2(input_path, output_path)
        
        return {
            'success': True,
            'output': str(output_path),
            'size': output_path.stat().st_size,
            'note': '演示模式 - 实际转换需要安装转换库'
        }
    
    def find_libreoffice(self):
        """查找 LibreOffice"""
        paths = [
            r'C:\Program Files\LibreOffice\program\soffice.exe',
            r'C:\Program Files (x86)\LibreOffice\program\soffice.exe',
            '/usr/bin/libreoffice',
            '/usr/bin/soffice'
        ]
        
        for path in paths:
            if os.path.exists(path):
                return path
        return None
    
    def find_ffmpeg(self):
        """查找 FFmpeg"""
        # 检查 PATH 中是否有 ffmpeg
        ffmpeg_path = shutil.which('ffmpeg')
        if ffmpeg_path:
            return ffmpeg_path
        
        # 检查常见安装位置
        paths = [
            r'C:\ffmpeg\bin\ffmpeg.exe',
            r'C:\Program Files\ffmpeg\bin\ffmpeg.exe',
            r'C:\Users\%USERNAME%\ffmpeg\bin\ffmpeg.exe'
        ]
        
        for path in paths:
            path = os.path.expandvars(path)
            if os.path.exists(path):
                return path
        return None


# 批量转换
def batch_convert(folder_path, output_format, quality=90):
    """批量转换文件夹中的所有文件"""
    converter = FileConverter()
    results = {'success': 0, 'failed': 0, 'files': []}
    
    for root, dirs, files in os.walk(folder_path):
        for file in files:
            file_path = os.path.join(root, file)
            ext = os.path.splitext(file)[1][1:].lower()
            
            if converter.detect_file_type(ext):
                result = converter.convert(file_path, output_format, quality)
                if result['success']:
                    results['success'] += 1
                    results['files'].append(result['output'])
                else:
                    results['failed'] += 1
    
    return results


# 命令行使用
if __name__ == '__main__':
    import sys
    
    if len(sys.argv) < 3:
        print("用法：python converter.py <输入文件> <输出格式> [质量]")
        print("示例：python converter.py photo.jpg png 95")
        sys.exit(1)
    
    input_file = sys.argv[1]
    output_format = sys.argv[2]
    quality = int(sys.argv[3]) if len(sys.argv) > 3 else 90
    
    converter = FileConverter()
    result = converter.convert(input_file, output_format, quality)
    
    if result['success']:
        print(f"✅ 转换成功：{result['output']}")
        print(f"📊 文件大小：{result['size']} 字节")
    else:
        print(f"❌ 转换失败：{result['error']}")
