flipfile-box/requirements.txt
# GUI
tkinter

# 加密
cryptography

# 文件转换
Pillow>=9.0.0  # 图片转换

# 文档转换（可选）
# python-docx  # Word 文档
# reportlab    # PDF 生成
# pdf2docx     # PDF 转 Word

# 视频转换（可选）
# 需要单独安装 FFmpeg: https://ffmpeg.org/download.html

# 网络同步
requests

# 其他
pathlib
